requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.1
cvzone==1.4.1
cycler==0.12.1
distlib==0.3.9
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.1.24
fonttools==4.55.8
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mediapipe==0.10.20
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
scipy==1.15.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.31.2
Werkzeug==3.1.3
